+++ src/features/project/components/content/ProjectTask.jsx
@@ -1,77 +1,8 @@
-import { useContext, useEffect, useRef, useState } from "react";
-import AddOrEditProject from "./AddOrEditProject";
+import { useEffect, useRef } from "react";
 import { v4 as uuidv4 } from "uuid";
-import { ProjectContext } from "../context/ProjectContext";
-import Modal from "../../../components/Modal";
-import ProjectDeletionConfirmation from "./ProjectDeletionConfirmation";
-import ProjectStorage from "../data/ProjectStorage";
+import ProjectStorage from "../../data/ProjectStorage";
 
-const titleStyle = "text-3xl font-bold py-2 pt-4 sm:pt-6 md:pt-8";
-
-export default function ProjectContent() {
-  const {
-    items,
-    selected,
-    handleSaveEdit,
-    handleDelete,
-    handleUpdateOrCreate,
-  } = useContext(ProjectContext);
-  const [showModal, setShowModal] = useState(false);
-
-  if (selected.isEditing) return <AddOrEditProject />;
-
-  return (
-    <div className="in-slide-down-realfast">
-      <Modal open={showModal}>
-        <ProjectDeletionConfirmation
-          onDelete={() => handleDelete(selected.item)}
-          showModal={setShowModal}
-        />
-      </Modal>
-
-      <div align="right">
-        <button
-          type="button"
-          className="btn-secondary-alert mx-1"
-          onClick={() => setShowModal(true)}
-        >
-          Delete
-        </button>
-        <button
-          type="submit"
-          className="btn-primary mx-1"
-          onClick={() => handleUpdateOrCreate(selected.item)}
-        >
-          Edit
-        </button>
-      </div>
-
-      <ProjectContentMain project={selected} />
-
-      <hr className="opacity-20 border-1 my-4 border-green-800" />
-
-      <ProjectContentTask
-        items={items}
-        project={selected}
-        updateUI={handleSaveEdit}
-      />
-    </div>
-  );
-}
-
-export function ProjectContentMain({ project }) {
-  return (
-    <div className="overflow-x-auto custom-scrollbar">
-      <h1 className={titleStyle}>{project.item.title}</h1>
-      <p className="text-gray-500">{project.item.dueDate}</p>
-      <p className={`py-4 ${project.item.desc || "text-gray-500"}`}>
-        {project.item.desc || "No description"}
-      </p>
-    </div>
-  );
-}
-
-export function ProjectContentTask({ items, project, updateUI }) {
+export default function ProjectTask({ items, project, updateUI }) {
   const inputStyle =
     "input-field mr-2 task-list opacity-85 hover:!opacity-100 focus:!opacity-100";
 
@@ -96,7 +27,6 @@ export function ProjectContentTask({ items, project, updateUI }) {
 
   return (
     <>
-      <h1 className={titleStyle}>Tasks</h1>
       <div className="my-2">
         <input
           type="text"
