+++ src/routes/router.jsx
@@ -1,0 +1,5 @@
+import { createBrowserRouter } from "react-router";
+
+const router = createBrowserRouter([]);
+
+export default router;
